requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.134.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-genai==1.65.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
multidict==6.7.1
numpy==2.4.2
orjson==3.11.7
packaging==24.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pinecone==8.1.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
vertexai==1.71.1
websockets==16.0
yarl==1.22.0